Beginning the move over to a new system for storing and managing raids, which will be more robust and human-readable than bloated XML.
