+++ Core/BotRaidCommands.cs
@@ -229,302 +229,12 @@ namespace DiscordBot.Core
             }
         }
 
-        /* Spooky unsafe C stuff */
+        /* Spooky unsafe C++ stuff */
 
         [DllImport("libherrington")]
         private static extern void solve(uint idx, IntPtr users, int length, IntPtr output);
 
-        /* End of spooky unsafe C stuff */
-
-        struct Raid
-        {
-            public string fullName;
-            public long   timestamp;
-            public int    raidID;
-        };
-
-        /// <summary>
-        /// Enumerates all the raids
-        /// </summary>
-        /// <returns>Returns an enumerable collection of Raids.</returns>
-        private static IEnumerable<Raid> EnumerateRaids()
-        {
-            //Create a regex to match a raid folder
-            Regex regex = new Regex(@"^raid_(\d+)_(\d+)$");
-
-            //Return a collection of raids
-            return Directory.EnumerateDirectories("./raids/") //Enumerate folders under raid directory
-                            .Where (f => regex.IsMatch(f))    //Filter out non-raids
-                            .Select(f => regex.Match(f))      //Extract timestamp and ID
-                            .Select(r => new Raid             //Wrap in a simple aggregate object  
-                            {
-                                fullName  = r.Name,
-                                timestamp = long.Parse(r.Groups[1].Value),
-                                raidID    =  int.Parse(r.Groups[2].Value)
-                            });
-        }
-
-        /// <summary>
-        /// Searches for an available ID to assign to a new raid.
-        /// </summary>
-        /// <returns>
-        /// A number between 1 and <see cref="Int32.MaxValue"/> if successful,
-        /// returns -1 if the operation failed.
-        /// </returns>
-        private static int GetNextAvailableRaidID()
-        {
-            //Create a sorted set to hold used IDs
-            SortedSet<int> usedIDs = new SortedSet<int>();
-
-            //Catch any errors
-            try
-            {
-                //Find used raid IDs
-                var coll = EnumerateRaids().Select(r => r.raidID);
-
-                //Iterate over ids
-                foreach (int id in coll)
-                {
-                    //Add to the sorted set
-                    usedIDs.Add(id);
-                }
-
-                //If collection is empty we just return 1
-                if (usedIDs.Count == 0) return 1;
-
-                //Starting at 1, loop until we find an unused id
-                for (int i = 1; i < usedIDs.Max; i++)
-                {
-                    //Check if the set does not contain the ID
-                    if (!usedIDs.Contains(i))
-                    {
-                        //Return this ID
-                        return i;
-                    }
-                }
-
-                //Return Max + 1
-                return usedIDs.Max + 1;
-            }
-            catch (Exception ex)
-            {
-                //Log error
-                Logger.Log(LOG_LEVEL.ERROR, ex.Message);
-
-                //Return failure
-                return -1;
-            }
-        }
-
-        /// <summary>
-        /// Initialises a new raid.
-        /// </summary>
-        /// <param name="offset">The time offset for when the raid starts.</param>
-        /// <param name="description">The description of the raid.</param>
-        /// <returns>
-        /// Returns the assigned ID for the raid which will be in the range 1 to
-        /// <see cref="Int32.MaxValue"/> if successful, otherwise it will return -1.
-        /// </returns>
-        private static int CreateRaid(DateTimeOffset offset, string description)
-        {
-            //Find an available ID
-            int raidID = GetNextAvailableRaidID();
-
-            //Catch any exceptions
-            try
-            {
-                //Check that the ID is valid
-                Debug.Assert(raidID > 0, "Couldn't get a valid raid ID");
-
-                //Determine the folder name
-                var name = $"raid_{offset.ToUnixTimeSeconds()}_{raidID}";
-
-                //Create directory for the raid
-                var dir = Directory.CreateDirectory($"./raids/{name}/");
-
-                //Check that it was created
-                Debug.Assert(dir.Exists, "Couldn't create raid directory");
-
-                //Create the initial roster file
-                using (FileStream fs = File.Open($@"./raids/{name}/roster.txt", FileMode.CreateNew, FileAccess.Write, FileShare.None))
-                {
-                    //Get a UTF8 encoded text stream
-                    using (StreamWriter sw = new StreamWriter(fs, Encoding.UTF8))
-                    {
-                        //Write description
-                        sw.WriteLine(description);
-                    }
-                }
-
-                //Return success
-                return raidID;
-            }
-            catch (Exception ex)
-            {
-                //Log error
-                Logger.Log(LOG_LEVEL.ERROR, ex.Message);
-
-                //Return failure
-                return -1;
-            }
-        }
-
-        /// <summary>
-        /// Appends a raider to the raid roster.
-        /// Note: Changes to role will simply be documented with a later entry to the file.
-        /// Shouldn't be an issue.
-        /// </summary>
-        /// <param name="raidID">The ID of the raid.</param>
-        /// <param name="userID">The user's unique ID on Discord.</param>
-        /// <param name="roles">A space-separated list of roles in upper-case.</param>
-        /// <returns>Returns whether the operation was successful or not.</returns>
-        private static bool AppendRaider(int raidID, UInt64 userID, string roles)
-        {
-            //Catch any errors
-            try
-            {
-                //Find the correct raid folder
-                var raid = EnumerateRaids().Where(r => r.raidID == raidID).First();
-
-                //Keep trying until we succeed
-                Utility.WithRetry((retryNum) =>
-                {
-                    //Sleep between retries
-                    if (retryNum > 0) Thread.Sleep(100);
-
-                    //Catch IO errors thrown when we can't get exclusive access
-                    try
-                    {
-                        //Get an exclusive handle to the file
-                        using (FileStream fs = File.Open($@"./raids/{raid.fullName}/roster.txt", FileMode.Append, FileAccess.Write, FileShare.None))
-                        {
-                            //Get a UTF8 encoded text stream
-                            using (StreamWriter sw = new StreamWriter(fs, Encoding.UTF8))
-                            {
-                                //Append raider
-                                sw.WriteLine($"{userID} - {roles}");
-                            }
-                        }
-
-                        //Return success
-                        return true;
-                    }
-                    catch (IOException ex)
-                    {
-                        //Check that it's not some other error
-                        Debug.Assert(ex.HResult == -2147024864, "Unexpected IO exception.");
-
-                        //Return failure
-                        return false;
-                    }
-                }, 10);
-
-                //Return success
-                return true;
-            }
-            catch (Exception ex)
-            {
-                //Log error
-                Logger.Log(LOG_LEVEL.ERROR, ex.Message);
-
-                //Return failure
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Reads the generated raid composition for a given raid.
-        /// </summary>
-        /// <param name="raidID">The ID of the raid.</param>
-        /// <returns>An array containing the user IDs of the people in the raid.</returns>
-        private static UInt64[] ReadRaidComp(int raidID)
-        {
-            //Catch any errors
-            try
-            {
-                //Find the correct raid folder
-                var raid = EnumerateRaids().Where(r => r.raidID == raidID).First();
-
-                //Prepare the string to hold the contents
-                string s = "";
-
-                //Keep trying until we succeed
-                Utility.WithRetry((retryNum) =>
-                {
-                    //Sleep between retries
-                    if (retryNum > 0) Thread.Sleep(100);
-
-                    //Catch IO errors thrown when we can't get exclusive access
-                    try
-                    {
-                        //Get an exclusive handle to the file
-                        using (FileStream fs = File.Open($@"./raids/{raid.fullName}/comp.txt", FileMode.Open, FileAccess.Read, FileShare.None))
-                        {
-                            //Get a UTF8 encoded text stream
-                            using (StreamReader sr = new StreamReader(fs, Encoding.UTF8))
-                            {
-                                //Read everything in the file
-                                s = sr.ReadToEnd();
-                            }
-                        }
-
-                        //Return success
-                        return true;
-                    }
-                    catch (IOException ex)
-                    {
-                        //Check that it's not some other error
-                        Debug.Assert(ex.HResult == -2147024864, "Unexpected IO exception.");
-
-                        //Return failure
-                        return false;
-                    }
-                }, 10);
-
-                //Return composition
-                return Utility.GetLines(s)                         //Extract lines
-                              .Select  (line => ulong.Parse(line)) //Parse each line
-                              .ToArray ();                         //Get array;
-            }
-            catch (Exception ex)
-            {
-                //Log error
-                Logger.Log(LOG_LEVEL.ERROR, ex.Message);
-
-                //Return failure
-                return null;
-            }
-        }
-
-        /// <summary>
-        /// Removes old raid files.
-        /// </summary>
-        /// <param name="maxAge">
-        /// The allowed age before the file is removed.
-        /// Starts counting after start of raid.
-        /// </param>
-        private static void CleanRaidFiles(TimeSpan maxAge)
-        {
-            //Get current time in UTC+0
-            var now = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
-
-            //Find the raids we want to delete
-            EnumerateRaids().Where(r =>
-            {
-                //Calculate age in seconds
-                var age = (now - r.timestamp);
-
-                //Wrap age in a TimeSpan object
-                var t = TimeSpan.FromTicks(TimeSpan.TicksPerSecond * age);
-
-                //Keep only old raids
-                return t > maxAge;
-            }).ToList().ForEach(r =>
-            {
-                //Try deleting directory (with contents)
-                Debug.Try(() => Directory.Delete($@"./raids/{r.fullName}", true));
-            });
-        }
+        /* End of spooky unsafe C++ stuff */
 
         /* Nice C# wrapper for spooky unsafe C++ stuff */
